decomp: sweep-1 harvest — 13 band3 functions to 100% via grind + close-out
First production grind sweep (113 fns x 3 passes, $0.51): 2 outright sweep
matches (CalibrationPanel::UpdateStream 60->100 by dropping a dev-only branch
retail omits; AccomplishmentPlayerConditional::IsConditionMet 94->100) + 11
closed from >=90% drafts by the close-out wave (local extraction for stack
slots/regalloc, signed-vs-unsigned compares, bool-mask forms, retail-vs-Wii
constant SaveSize*19, decl-order vs static guards, reciprocal-multiply float
form). Whole-binary: 11278 -> 11291 (+13), 0 regressions (per-group worktree
A/B gates + report.json verification on main).

+++ src/band3/game/TrainerPanel.cpp
@@ -160,7 +160,7 @@ int TrainerPanel::FixupTick(int tick) const {
 
 void TrainerPanel::ResetChallenge() {
     TrainerChallenge *tc = mSections[mCurrSection].mChallenge;
-    if (tc) {
+    if ((int)tc) {
         tc->Exit();
         mSections[mCurrSection].mChallenge->Enter();
     }
